# AplicationDatariumApi/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "AplicationDatariumApi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Indicators/selic": {
      "get": {
        "tags": [
          "Indicators"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "type": "string",
            "format": "date-time"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Indicators/selic/history": {
      "get": {
        "tags": [
          "Indicators"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "start",
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "query",
            "name": "end",
            "type": "string",
            "format": "date-time"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
...
}